requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-comments-xtd==2.9.2
django-contrib-comments==2.1.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
docutils==0.17.1
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
phonenumberslite==8.12.27
Pillow==8.3.1
psycopg2-binary==2.9.1
pyotp==2.6.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
